test: why relative path /assets/.../web.css breaks
Changing back from using /Assets/.../web.css in root/index.html

The original change was made because the folder for assets in GitHub
was named "Assets" while the local version in VSCode is "assets", so I
wanted to see if doing it with "Assets" would work in the GitHub hosted
version even if it would break the local versions (Jekyll/LiveServer).
It did, which means I have to clean up the two versions now.

Was originally confused because VSCode says there are no version
control issues.