fix: commit w/ "git config core.ignorecase false"
A major problem that I encountered recently was my local href paths
breaking after I changed my naming conventions in my local file
structures because I learned that for frontend dev the general best
practice is to have everything in lower case. However, my git was
configured to be case insensitive which meant that when I changed my
"Assets" folder to "assets" the change was only properly reflected in
VSCode and not GitHub (which is the most important because it's
literally where this page is hosted). So hopefully this commit will fix
this issue and restore the functional paths.